Kuba choose, edit and create profile

+++ User.cs
@@ -17,10 +17,10 @@ namespace app
 		public string name { get; set; }
 		public byte age { get; set; }
 		public float weight { get; set; }
-		public float height { get; set; }
+		public uint height { get; set; }
 		public Gender gender { get; set; }
 
-		public User(string _name, byte _age, float _weight, float _height, Gender _gender)
+		public User(string _name, byte _age, float _weight, uint _height, Gender _gender)
 		{
 			name = _name;
 			age = _age;
